+++ meja/src/main/java/com/dua3/meja/model/AbstractRow.java
@@ -1,5 +1,7 @@
 package com.dua3.meja.model;
 
+import java.util.Optional;
+
 /**
  * Abstract base class for implementations of the {@link Row} interface.
  */
@@ -17,7 +19,7 @@ public abstract class AbstractRow implements Row {
     public abstract AbstractCell getCell(int col);
 
     @Override
-    public abstract AbstractCell getCellIfExists(int j);
+    public abstract Optional<? extends AbstractCell> getCellIfExists(int j);
 
     @Override
     public int getRowNumber() {